aquasense-ml/requirements.txt
fastapi
uvicorn
scikit-learn
pandas
joblib
pymongo
dnspython
python-dotenv
